requirements.txt
autogen-agentchat>=0.4
autogen-ext[openai,mcp]>=0.4
autogen-core>=0.4
Pillow>=10.0
scikit-image>=0.22
numpy>=1.26
fastapi>=0.115
uvicorn[standard]>=0.32
websockets>=13.0
python-dotenv>=1.0
